Moved SurveyFormState under webapp/store/ui/surveyForm
