+++ webapp/loggedin/surveyViews/reducer.js
@@ -1,8 +1,7 @@
 import { combineReducers } from 'redux'
 
 import { RecordReducer, RecordState } from '@webapp/store/ui/record'
-import { SurveyFormReducer } from '@webapp/store/ui/surveyForm'
-import * as SurveyFormState from './surveyForm/surveyFormState'
+import { SurveyFormReducer, SurveyFormState } from '@webapp/store/ui/surveyForm'
 
 export default combineReducers({
   [RecordState.stateKey]: RecordReducer,
