+++ webapp/store/ui/surveyForm/index.js
@@ -1,4 +1,5 @@
 import * as SurveyFormActions from './actions'
 import SurveyFormReducer from './reducer'
+import * as SurveyFormState from './state'
 
-export { SurveyFormActions, SurveyFormReducer }
+export { SurveyFormActions, SurveyFormReducer, SurveyFormState }
